requirements.txt
Django
psycopg2
pytz
djangorestframework
markdown
django-filter